# LibGfx: Decode Mellor_ACTITC_10.pdf faster [3/N]