Related to #26082

Before:
``` bash
Benchmark 1: BuildLagom/bin/pdf --render out.png PDF/Mellor_ACTITC_10.pdf --page 4
  Time (mean ± σ):     427.7 ms ±   8.4 ms    [User: 381.2 ms, System: 46.3 ms]
  Range (min … max):   410.3 ms … 438.6 ms    10 runs
```

After:
``` bash
Benchmark 1: BuildLagom/bin/pdf --render out.png PDF/Mellor_ACTITC_10.pdf --page 4
  Time (mean ± σ):     337.2 ms ±   3.0 ms    [User: 295.2 ms, System: 41.8 ms]
  Range (min … max):   332.0 ms … 340.7 ms    10 runs
```